Organizes compoment folders by moving into apprioriate folder and updating import file locations for pages
